refactor: has_contract_been_built function
